Removed public setter visibility with type interfaces for each of the model classes
