+++ src/main/java/org/jboss/jdf/stacks/model/BomVersionImpl.java
@@ -16,23 +16,17 @@
  */
 package org.jboss.jdf.stacks.model;
 
-public class ArchetypeImpl implements Archetype {
+import java.util.Properties;
+
+public class BomVersionImpl implements BomVersion{
 
     private String id;
 
-    private String name;
+    private BomImpl bom;
 
-    private String description;
+    private String version;
 
-    private String groupId;
-
-    private String artifactId;
-
-    private String recommendedVersion;
-    
-    private String repositoryURL;
-
-    private Archetype blank;
+    private Properties labels = new Properties();
 
     public String getId() {
         return id;
@@ -42,60 +36,28 @@ public class ArchetypeImpl implements Archetype {
         this.id = id;
     }
 
-    public String getName() {
-        return name;
+    public BomImpl getBom() {
+        return bom;
     }
 
-    public void setName(String name) {
-        this.name = name;
+    public void setBom(BomImpl bom) {
+        this.bom = bom;
     }
 
-    public String getDescription() {
-        return description;
+    public String getVersion() {
+        return version;
     }
 
-    public void setDescription(String description) {
-        this.description = description;
+    public void setVersion(String version) {
+        this.version = version;
     }
 
-    public String getGroupId() {
-        return groupId;
+    public Properties getLabels() {
+        return labels;
     }
 
-    public void setGroupId(String groupId) {
-        this.groupId = groupId;
-    }
-
-    public String getArtifactId() {
-        return artifactId;
-    }
-
-    public void setArtifactId(String artifactId) {
-        this.artifactId = artifactId;
-    }
-
-    public String getRecommendedVersion() {
-        return recommendedVersion;
-    }
-
-    public void setRecommendedVersion(String recommendedVersion) {
-        this.recommendedVersion = recommendedVersion;
-    }
-    
-    public String getRepositoryURL() {
-        return repositoryURL;
-    }
-    
-    public void setRepositoryURL(String repositoryURL) {
-        this.repositoryURL = repositoryURL;
-    }
-
-    public Archetype getBlank() {
-        return blank;
-    }
-
-    public void setBlank(Archetype blank) {
-        this.blank = blank;
+    public void setLabels(Properties labels) {
+        this.labels = labels;
     }
 
     /*
@@ -124,10 +86,10 @@ public class ArchetypeImpl implements Archetype {
         if (obj == null) {
             return false;
         }
-        if (!(obj instanceof ArchetypeImpl)) {
+        if (!(obj instanceof BomVersionImpl)) {
             return false;
         }
-        ArchetypeImpl other = (ArchetypeImpl) obj;
+        BomVersionImpl other = (BomVersionImpl) obj;
         if (id == null) {
             if (other.id != null) {
                 return false;
@@ -145,7 +107,7 @@ public class ArchetypeImpl implements Archetype {
      */
     @Override
     public String toString() {
-        return "Archetype [id=" + id + "]";
+        return "BomVersion [id=" + id + "]";
     }
 
 }
